[BREAKING] Move functionality to extension packages (#328)
* move jump to extension

* support for regular dependency in julia 1.6

* reorganize JuMP ext

* move RecipesBase and GeometryBasics to ext

* uncomment verbose statement

* add error message for Mesh

* add news file describing breaking change

* update docs

* update examples

* simplify doc

* fix verbose print of model

* remove exported JuMP function

* import functions missing in GeometryBasicsExt

* add missing compat entry
